+++ app/controllers/image-sets/index.js
@@ -8,8 +8,15 @@ export default Ember.Controller.extend({
   selectedOrganization: Ember.computed.reads('currentOrganization'),
 
   canView: Ember.computed.alias('activeImageSet'),
-  canRequestCv: Ember.computed.alias('activeImageSet'),
   canViewCv: Ember.computed.alias('activeImageSet.hasCvResults'),
+  cvRequestPending: Ember.computed.alias('activeImageSet.cvRequestPending'),
+
+  canRequestCv: function() {
+    var activeImageSet = this.get('activeImageSet');
+
+    return activeImageSet && !activeImageSet.get('hasCvRequest');
+  }.property('activeImageSet', 'activeImageSet.hasCvRequest'),
+
   canDelete: function() {
     return this.get('activeImageSet.uploadingOrganization') === this.get('currentOrganization');
   }.property('activeImageSet.uploadingOrganization', 'currentOrganization'),
@@ -35,8 +42,10 @@ export default Ember.Controller.extend({
           canDelete = this.get('canDelete');
 
       if (activeImageSet && canDelete) {
-        this.set('activeImageSet', null);
-        activeImageSet.destroyRecord();
+        if (confirm("Are you sure you want to delete image set " + activeImageSet.get('id') + "? this cannot be undone.")) {
+          this.set('activeImageSet', null);
+          activeImageSet.destroyRecord();
+        }
       }
     },
